move most of the heavy lifting to AddResponder
It makes sense to extract all the necessary information from the responder
while it's being added; this improves performance and error handling (e.g.
if there is a mismatch between the Path pattern in an EndpointAttribute and
the parameters a method actually provides, an exception can be thrown while
the responder is being added). This also accelerates matching a request
against a handler method.